first working refactor
