+++ commands/fun/8ball.js
@@ -1,6 +1,6 @@
 const Discord = require('discord.js');
 const config = require('../../configuration/config.json');
-const winstonLogHandler = require('../../handler/winstonLogHandler');
+const winstonLogHandler = require('../../handler/util/winstonLogHandler');
 const logger = winstonLogHandler.getLogger();
 
 module.exports = {
