+++ commands/league/addleagueaccount.js
@@ -1,6 +1,6 @@
 const Discord = require('discord.js');
-const mariadbHandler = require('../../handler/mariadbHandler');
-const winstonLogHandler = require('../../handler/winstonLogHandler');
+const mariadbHandler = require('../../handler/util/mariadbHandler');
+const winstonLogHandler = require('../../handler/util/winstonLogHandler');
 const apiKeys = require('../../configuration/apiKeyConfig');
 const lolApi = require('league-api-2.0');
 const logger = winstonLogHandler.getLogger();
