+++ commands/league/lolstatus.js
@@ -2,7 +2,7 @@ const Discord = require('discord.js');
 const lolApi = require('league-api-2.0');
 const config = require('../../configuration/config');
 const apiKeys = require('../../configuration/apiKeyConfig');
-const winstonLogHandler = require('../../handler/winstonLogHandler');
+const winstonLogHandler = require('../../handler/util/winstonLogHandler');
 const logger = winstonLogHandler.getLogger();
 
 module.exports = {
