+++ commands/moderation/setGuildPrefix.js
@@ -1,5 +1,5 @@
-const winstonLogHandler = require('../../handler/winstonLogHandler');
-const mariadbHandler = require('../../handler/mariadbHandler');
+const winstonLogHandler = require('../../handler/util/winstonLogHandler');
+const mariadbHandler = require('../../handler/util/mariadbHandler');
 const logger = winstonLogHandler.getLogger();
 
 module.exports = {
