+++ commands/moderation/showGuildPrefix.js
@@ -1,5 +1,5 @@
-const winstonLogHandler = require('../../handler/winstonLogHandler');
-const mariadbHandler = require('../../handler/mariadbHandler');
+const winstonLogHandler = require('../../handler/util/winstonLogHandler');
+const mariadbHandler = require('../../handler/util/mariadbHandler');
 const config = require('../../configuration/config');
 const logger = winstonLogHandler.getLogger();
