+++ events/ready.js
@@ -1,4 +1,4 @@
-const winstonLogHandler = require('../handler/winstonLogHandler');
+const winstonLogHandler = require('../handler/util/winstonLogHandler');
 const logger = winstonLogHandler.getLogger();
 const config = require('../configuration/config');
