+++ handler/util/discordEventHandler.js
@@ -6,7 +6,7 @@ exports.run = (client) => {
     fs.readdir('./events/', (err, files) => {
         if (err) return console.error(err);
         files.forEach(file => {
-            const eventFunction = require(`./../events/${file}`);
+            const eventFunction = require(`./../../events/${file}`);
             const eventName = file.split('.')[0];
             client.on(eventName, (...args) => eventFunction.run(client, ...args));
         });
